+++ app/Http/Controllers/NivelEnsinoController.php
@@ -3,6 +3,7 @@
 namespace SerEducacional\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Prettus\Validator\Contracts\ValidatorInterface;
 use Prettus\Validator\Exceptions\ValidatorException;
 use SerEducacional\Repositories\NivelEnsinoRepository;
@@ -89,15 +90,22 @@ class NivelEnsinoController extends Controller
 
         #Editando a grid
         return Datatables::of($rows)->addColumn('action', function ($row) {
+            # Recuperando o usuário
+            $user = Auth::user();
 
             #verificando se existe vinculo com outra tabela (servidores)
             $curso = $this->cursoRepository->findWhere(['nivel_ensino_id' => $row->id]);
 
-            #botão editar
-            $html  = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i></a> ';
+            #html de uso
+            $html = '';
+
+            # Verificando a permissão de editar
+            if($user->can('nivel.update')) {
+                $html  = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i></a> ';
+            }
 
             #condição para que habilite a opção de remover
-            if (count($curso) == 0) {
+            if (count($curso) == 0 && $user->can('nivel.destroy')) {
                 $html .= '<a href="destroy/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-remove"></i></a>';
             }
