+++ app/Http/Controllers/PessoaJuridicaController.php
@@ -4,6 +4,7 @@ namespace SerEducacional\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Auth;
 use SerEducacional\Http\Requests;
 use Prettus\Validator\Contracts\ValidatorInterface;
 use Prettus\Validator\Exceptions\ValidatorException;
@@ -108,8 +109,21 @@ class PessoaJuridicaController extends Controller
         
         #Editando a grid
         return Datatables::of($rows)->addColumn('action', function ($row) {
-            $html  = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i></a> ';
-            $html .= '<a href="destroy/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-remove"></i></a>';
+            # Recuperando o usuário
+            $user = Auth::user();
+
+            # Variáveis de uso
+            $html  = '';
+
+            # Verificando a permissão de editar
+            if($user->can('pessoa.juridica.update')) {
+                $html  = '<a href="edit/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-edit"></i></a> ';
+            }
+
+            # Verificando a permissão de remorção
+            if($user->can('pessoa.juridica.destroy')) {
+                $html .= '<a href="destroy/'.$row->id.'" class="btn btn-xs btn-primary"><i class="glyphicon glyphicon-remove"></i></a>';
+            }
 
             # Retorno
             return $html;
